% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{plot the popularity of babynames in times}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{name in the dataset}

\item{the_sex}{sex of the chosen name}
}
\value{
graph
}
\description{
plot the popularity of babynames in times
}
\examples{
\dontrun{
draw_a_name("...","M")

}

}
