% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{plot simultaneously the popularity of more than one name}
\usage{
draw_names(the_names)
}
\arguments{
\item{the_names}{name in the dataset}
}
\value{
graph
}
\description{
plot simultaneously the popularity of more than one name
}
\examples{
\dontrun{
draw_names(c("Diane"."Vincent"))

}


}
