% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{multi_excel}
\alias{multi_excel}
\title{import all sheets from an xlsx file}
\usage{
multi_excel(file)
}
\arguments{
\item{file}{path to a an xlsx file}
}
\value{
list
}
\description{
import all sheets from an xlsx file
}
\examples{
\dontrun{
multi_excel("../../name.xlsx")

}



}
