% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{import all csv from a folder}
\usage{
my_csv_reader(folder)
}
\arguments{
\item{folder}{path to the folder}
}
\value{
a list
}
\description{
import all csv from a folder
}
\examples{
\dontrun{

my_csv_reader(../../../path)

}

}
