% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{save a data.frame in a csv file}
\usage{
save_as_csv(dataset, filename, row.names = FALSE)
}
\arguments{
\item{dataset}{a data.frame}

\item{filename}{the name of the output csv}

\item{row.names}{attribute of the data.frame}
}
\value{
csv file
}
\description{
save a data.frame in a csv file
}
\examples{
\dontrun{
save_as_csv(..,"name.csv")

}


}
