% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{multi_excel}
\alias{multi_excel}
\title{Multiple excel sheets}
\usage{
multi_excel(excel_file_name)
}
\arguments{
\item{file_name}{full path including file name}
}
\value{
list dataframes of excel sheets of a xlsx file
}
\description{
Multiple excel sheets
}
\examples{
\dontrun{
 multi_excel("..")
}
}
