% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{read csv files}
\usage{
my_csv_reader(dir_path = getwd())
}
\arguments{
\item{dir_path}{directory from which the csv files are read}
}
\value{
list of csv files
}
\description{
read csv files
}
\examples{
\dontrun{
my_csv_reader("..")
}
}
